Measure Fibonacci Performance.
In this commit, the basic fibonacci functions are changed so that
the number of recursive function calls can be measured. This is
done by surrounding the function body with a closure and introducing
a variable that keeps track of function calls. Also a more efficient
version of the fibonacci function is introduced that calculates the
result iteratively using a loop. This executes in linear time.